+++ source/op/prod_force_grad_multi_device.cc
@@ -107,11 +107,14 @@ class ProdForceSeAGradOp : public OpKernel {
     assert(nframes == net_deriv_tensor.shape().dim_size(0));
     assert(nframes == in_deriv_tensor.shape().dim_size(0));
     assert(nframes == nlist_tensor.shape().dim_size(0));
-    assert(nloc * ndescrpt == grad_net_shape.dim_size(1));
+    assert(static_cast<int64_t>(nloc) * ndescrpt == grad_net_shape.dim_size(1));
     assert(nloc * 3 == grad_shape.dim_size(1));
-    assert(nloc * ndescrpt == net_deriv_tensor.shape().dim_size(1));
-    assert(nloc * ndescrpt * 3 == in_deriv_tensor.shape().dim_size(1));
-    assert(nloc * nnei == nlist_tensor.shape().dim_size(1));
+    assert(static_cast<int64_t>(nloc) * ndescrpt ==
+           net_deriv_tensor.shape().dim_size(1));
+    assert(static_cast<int64_t>(nloc) * ndescrpt * 3 ==
+           in_deriv_tensor.shape().dim_size(1));
+    assert(static_cast<int64_t>(nloc) * nnei ==
+           nlist_tensor.shape().dim_size(1));
     assert(nnei * 4 == ndescrpt);
     // flat the tensors
     FPTYPE* p_grad_net = grad_net_tensor->flat<FPTYPE>().data();
@@ -215,11 +218,14 @@ class ProdForceSeRGradOp : public OpKernel {
     assert(nframes == net_deriv_tensor.shape().dim_size(0));
     assert(nframes == in_deriv_tensor.shape().dim_size(0));
     assert(nframes == nlist_tensor.shape().dim_size(0));
-    assert(nloc * ndescrpt == grad_net_shape.dim_size(1));
+    assert(static_cast<int64_t>(nloc) * ndescrpt == grad_net_shape.dim_size(1));
     assert(nloc * 3 == grad_shape.dim_size(1));
-    assert(nloc * ndescrpt == net_deriv_tensor.shape().dim_size(1));
-    assert(nloc * ndescrpt * 3 == in_deriv_tensor.shape().dim_size(1));
-    assert(nloc * nnei == nlist_tensor.shape().dim_size(1));
+    assert(static_cast<int64_t>(nloc) * ndescrpt ==
+           net_deriv_tensor.shape().dim_size(1));
+    assert(static_cast<int64_t>(nloc) * ndescrpt * 3 ==
+           in_deriv_tensor.shape().dim_size(1));
+    assert(static_cast<int64_t>(nloc) * nnei ==
+           nlist_tensor.shape().dim_size(1));
     assert(nnei * 1 == ndescrpt);
     // flat the tensors
     FPTYPE* p_grad_net = grad_net_tensor->flat<FPTYPE>().data();
